+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/GoneException.java
@@ -6,7 +6,6 @@ package com.azure.cosmos.implementation;
 import com.azure.cosmos.BridgeInternal;
 import com.azure.cosmos.CosmosException;
 import com.azure.cosmos.implementation.directconnectivity.HttpUtils;
-import com.azure.cosmos.implementation.directconnectivity.Uri;
 import com.azure.cosmos.implementation.http.HttpHeaders;
 
 import java.net.SocketAddress;