+++ sdk/cosmos/azure-cosmos/src/main/java/com/azure/cosmos/implementation/directconnectivity/rntbd/RntbdRequestManager.java
@@ -768,7 +768,7 @@ public final class RntbdRequestManager implements ChannelHandler, ChannelInbound
         for (RntbdRequestRecord record : this.pendingRequests.values()) {
 
             final Map<String, String> requestHeaders = record.args().serviceRequest().getHeaders();
-            final String requestUri = record.args().physicalAddressUri().toString();
+            final String requestUri = record.args().physicalAddressUri().getURI().toString();
 
             final GoneException error = new GoneException(message, cause, null, requestUri);
             BridgeInternal.setRequestHeaders(error, requestHeaders);
@@ -844,7 +844,7 @@ public final class RntbdRequestManager implements ChannelHandler, ChannelInbound
             // ..Create CosmosException based on status and sub-status codes
 
             final String resourceAddress = requestRecord.args().physicalAddressUri() != null ?
-                requestRecord.args().physicalAddressUri().toString() : null;
+                requestRecord.args().physicalAddressUri().getURI().toString() : null;
 
             switch (status.code()) {